lsm-updates/libs/lsm_update/requirements.dev.txt
inmanta-dev-dependencies[module]
pytest-inmanta-lsm